# docs/api-spec.yaml
openapi: 3.0.1
info:
  title: 항해 플러스 E-커머스
  description: ''
  version: 1.0.0
...
paths:
  /balances/{customerId}:
    get:
      summary: 사용자 잔액 조회
      deprecated: false
      description: ''
      tags:
        - Balance
      parameters:
        - name: customerId
          in: path
          description: 사용자 ID
          required: true
          example: '1'
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  customerId:
                    type: integer
                    description: 사용자 ID
                  amount:
                    type: integer
                    description: 사용자의 현재 잔액
                required:
                  - amount
                  - customerId
              example:
                customerId: 1
                amount: 10000
          headers: {}
      security: []
  /balances/charge:
    patch:
      summary: 사용자 잔액 충전
      deprecated: false
      description: ''
      tags:
        - Balance
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                custoemrId:
                  type: integer
                  description: 사용자 ID
                amount:
                  type: integer
                  description: 충전할 금액
              required:
                - amount
                - custoemrId
            example:
              customerId: 1
              amount: 50000
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  customerId:
                    type: integer
                    description: 사용자 ID
                  amount:
                    type: integer
                    description: 충전 후 잔액
                required:
                  - customerId
                  - amount
              example:
                customerId: 1
                amount: 150000
          headers: {}
      security: []
  /balances/{customerId}/balance-histories:
    get:
      summary: 사용자 잔액 변경 내역 조회
      deprecated: false
      description: ''
      tags:
        - Balance
      parameters:
        - name: customerId
          in: path
          description: 사용자 ID
          required: true
          example: '1'
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    changeType:
                      type: string
                      description: 잔액 변경 타입("CHARGE", "USE")
                    changeAmount:
                      type: integer
                      description: 변경된 액수
                    totalAmount:
                      type: integer
                      description: 변경 이후 총 잔액
                    createdAt:
                      type: string
                      description: 잔액 변경 시점
                  required:
                    - changeType
                    - changeAmount
                    - totalAmount
                    - createdAt
              example:
                - changeType: CHARGE
                  changeAmount: 10000
                  totalAmount: 60000
                  createdAt: '2025-04-02T16:31:11.959Z'
          headers: {}
      security: []
...
components:
...
  securitySchemes: {}
servers: []
security: []